Split logic for network chaos action into different module
